% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet.R
\name{BINNET}
\alias{BINNET}
\alias{report.BINNET}
\title{Neural Network Binary Time Series Forecasts}
\usage{
BINNET(formula, n_nodes = NULL, n_networks = 20, scale_inputs = TRUE, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section).}

\item{n_nodes}{Number of nodes in the hidden layer. Default is half of the
number of external regressors plus 1.}

\item{n_networks}{Number of networks to fit with different random starting
weights. These are then averaged when producing forecasts.}

\item{scale_inputs}{If TRUE, inputs are scaled by subtracting the column
means and dividing by their respective standard deviations.}

\item{...}{Other arguments passed to \verb{\link[nnet]\{nnet\}}.}
}
\value{
A model specification.
}
\description{
Feed-forward neural networks with a single hidden layer and lagged inputs
for forecasting univariate binary time series.
}
\details{
A feed-forward neural network is fitted with a single hidden layer containing \code{size} nodes.

Exogenous regressors are used as inputs.
A total of \code{repeats} networks are
fitted, each with random starting weights. These are then averaged when
computing forecasts.
}
\examples{
melb_rain |>
  model(nn = BINNET(Wet ~ fourier(K = 1, period = "year")))

}
