% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-fable-package.R
\docType{package}
\name{fable.binary-package}
\alias{fable.binary}
\alias{fable.binary-package}
\title{fable.binary: Forecasting Binary Time Series}
\description{
Provides a collection of time series forecasting models suitable for binary time series. These models work within the 'fable' framework provided by the 'fabletools' package, which provides the tools to evaluate, visualise, and combine models in a workflow consistent with the tidyverse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidyverts.github.io/fable.binary}
  \item Report bugs at \url{https://github.com/tidyverts/fable.binary/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} [copyright holder]

Authors:
\itemize{
  \item Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}
}

}
\keyword{internal}
