% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{fitted.LOGISTIC}
\alias{fitted.LOGISTIC}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{LOGISTIC}(object, ...)
}
\arguments{
\item{object}{Fitted model}

\item{...}{Other arguments ignored}
}
\description{
Extracts the fitted values.
}
\examples{
melb_rain |>
  model(logistic = LOGISTIC(Wet ~ fourier(K = 5, period = "year"))) |>
  fitted()
}
