% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet.R
\name{forecast.BINNET}
\alias{forecast.BINNET}
\title{Forecast a model from the fable package}
\usage{
\method{forecast}{BINNET}(object, new_data, specials = NULL, simulate = TRUE, times = 5000, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{simulate}{If \code{TRUE}, then forecast distributions are computed using simulation with Gaussian errors.}

\item{times}{The number of sample paths to use in estimating the forecast distribution when \code{bootstrap = TRUE}.}

\item{...}{Other arguments passed to methods}

\item{bootstrap}{If \code{TRUE}, then forecast distributions are computed using simulation with resampled errors.}
}
\description{
Produces forecasts from a trained model.
}
\examples{
as_tsibble(airmiles) \%>\%
  model(nn = BINNET(box_cox(value, 0.15))) \%>\%
  forecast(times = 10)
}
