% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet.R
\name{generate.BINNET}
\alias{generate.BINNET}
\title{Generate new data from a fable model}
\usage{
\method{generate}{BINNET}(x, new_data, specials = NULL, ...)
}
\arguments{
\item{x}{An object.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:generate.mdl_df]{fabletools::generate.mdl_df()}}).}

\item{...}{Other arguments passed to methods}
}
\description{
Simulates future paths from a dataset using a fitted model.
}
\examples{
melb_rain |>
  model(nn = BINNET(Wet ~ fourier(K = 1, period = "year"))) |>
  generate()
}
