% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{generate.LOGISTIC}
\alias{generate.LOGISTIC}
\title{Generate new data from a fable model}
\usage{
\method{generate}{LOGISTIC}(x, new_data, specials, bootstrap = FALSE, ...)
}
\arguments{
\item{x}{An object.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:generate.mdl_df]{fabletools::generate.mdl_df()}}).}

\item{bootstrap}{If \code{TRUE}, then innovations are sampled from the model's residuals.}

\item{...}{Other arguments passed to methods}
}
\description{
Simulates future paths from a dataset using a fitted model.
Innovations are sampled by the model's assumed error distribution.
If bootstrap is TRUE, innovations will be sampled from the model's residuals.
If new_data contains the .innov column, those values will be treated as innovations.
}
\examples{
as_tsibble(USAccDeaths) |>
  dplyr::mutate(y = value > mean(value)) |>
  model(logistic  = LOGISTIC(y ~ season())) \%>\%
  generate()
}
