% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{glance.LOGISTIC}
\alias{glance.LOGISTIC}
\title{Glance a LOGISTIC}
\usage{
\method{glance}{LOGISTIC}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\value{
A one row tibble summarising the model's fit.
}
\description{
Construct a single row summary of the LOGISTIC model.
}
\details{
Contains the R squared (\code{r_squared}), variance of residuals (\code{sigma2}),
the log-likelihood (\code{log_lik}), and information criterion (\code{AIC}, \code{AICc}, \code{BIC}).
}
\examples{
as_tsibble(USAccDeaths) |>
  dplyr::mutate(y = value > mean(value)) |>
  model(logistic  = LOGISTIC(y ~ season())) \%>\%
  glance()
}
