% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melb_rain}
\alias{melb_rain}
\title{Daily rainfall in Melbourne, Australia}
\format{
A \code{tsibble} data frame with 4322 rows and 3 variables.
}
\source{
Australian Bureau of Meteorology. \url{http://www.bom.gov.au/climate/data/}
}
\usage{
melb_rain
}
\description{
A dataset containing daily rainfall amounts from 1 January 2000 to 31 December 2011.
}
\keyword{datasets}
