% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet.R
\name{refit.BINNET}
\alias{refit.BINNET}
\title{Refit a BINNET model}
\usage{
\method{refit}{BINNET}(object, new_data, specials = NULL, reestimate = FALSE, ...)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{reestimate}{If \code{TRUE}, the networks will be initialized with random
starting weights to suit the new data. If \code{FALSE}, for every network the best
individual set of weights found in the pre-estimation process is used as the
starting weight vector.}

\item{...}{Other arguments passed to methods}
}
\value{
A refitted model.
}
\description{
Applies a fitted BINNET model to a new dataset.
}
