% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{refit.LOGISTIC}
\alias{refit.LOGISTIC}
\title{Refit a \code{LOGISTIC}}
\usage{
\method{refit}{LOGISTIC}(object, new_data, specials = NULL, reestimate = FALSE, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{new_data}{A tsibble containing the time points and exogenous regressors
for which a refit is required.}

\item{specials}{A list of special functions used in the model, (passed by
\code{fabletools::forecast.mdl_df}).}

\item{reestimate}{If TRUE, the networks will be initialized with random
starting weights to suit the new data. If FALSE, for every network the best
individual set of weights found in the pre-estimation process is used as the
starting weight vector.}

\item{...}{Other arguments passed to methods}
}
\description{
Applies a fitted \code{LOGISTIC} to a new dataset.
}
