% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{residuals.LOGISTIC}
\alias{residuals.LOGISTIC}
\title{Extract residuals from a fable model.}
\usage{
\method{residuals}{LOGISTIC}(object, ...)
}
\arguments{
\item{object}{Fitted model}

\item{...}{Other arguments ignored}
}
\description{
Extracts the residuals.
}
\examples{
as_tsibble(USAccDeaths) |>
  dplyr::mutate(y = value > mean(value)) |>
  model(logistic  = LOGISTIC(y ~ season())) \%>\%
  residuals()
}
