% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{tidy.LOGISTIC}
\alias{tidy.LOGISTIC}
\title{Tidy a fable model}
\usage{
\method{tidy}{LOGISTIC}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy \code{\link[tibble:tibble]{tibble::tibble()}}.}

\item{...}{Additional arguments to tidying method.}
}
\description{
Returns the coefficients from the model in a \code{tibble} format.
}
\examples{
melb_rain |>
  model(logistic = LOGISTIC(Wet ~ fourier(K = 5, period = "year"))) |>
  tidy()
}
