% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet.R
\name{fitted.BINNET}
\alias{fitted.BINNET}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{BINNET}(object, ...)
}
\arguments{
\item{object}{Fitted model}

\item{...}{Other arguments ignored}
}
\description{
Extracts the fitted values.
}
\examples{
as_tsibble(airmiles) \%>\%
  model(nn = BINNET(box_cox(value, 0.15))) \%>\%
  fitted()
}
