% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{fitted.LOGISTIC}
\alias{fitted.LOGISTIC}
\title{Extract fitted values from a fable model}
\usage{
\method{fitted}{LOGISTIC}(object, ...)
}
\arguments{
\item{object}{Fitted model}

\item{...}{Other arguments ignored}
}
\description{
Extracts the fitted values.
}
\examples{
as_tsibble(USAccDeaths) |>
  dplyr::mutate(y = value > mean(value)) |>
  model(logistic  = LOGISTIC(y ~ season())) \%>\%
  fitted()
}
