% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{forecast.LOGISTIC}
\alias{forecast.LOGISTIC}
\title{Forecast a model from the fable package}
\usage{
\method{forecast}{LOGISTIC}(
  object,
  new_data,
  specials = NULL,
  bootstrap = FALSE,
  approx_normal = TRUE,
  times = 5000,
  ...
)
}
\arguments{
\item{object}{A model for which forecasts are required.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:forecast]{fabletools::forecast.mdl_df()}}).}

\item{bootstrap}{If \code{TRUE}, then forecast distributions are computed using simulation with resampled errors.}

\item{approx_normal}{Should the resulting forecast distributions be
approximated as a Normal distribution instead of a Student's T
distribution. Returning Normal distributions (the default) is a useful
approximation to make it easier for using LOGISTIC models in model combinations
or reconciliation processes.}

\item{times}{The number of sample paths to use in estimating the forecast distribution when \code{bootstrap = TRUE}.}

\item{...}{Other arguments passed to methods}
}
\description{
Produces forecasts from a trained model.
}
\examples{
as_tsibble(USAccDeaths) |>
  dplyr::mutate(y = value > mean(value)) |>
  model(logistic  = LOGISTIC(y ~ season())) \%>\%
  forecast()
}
