% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{generate.LOGISTIC}
\alias{generate.LOGISTIC}
\title{Generate new data from a fable model}
\usage{
\method{generate}{LOGISTIC}(x, new_data, specials, ...)
}
\arguments{
\item{x}{An object.}

\item{new_data}{A tsibble containing the time points and exogenous regressors to produce forecasts for.}

\item{specials}{(passed by \code{\link[fabletools:generate.mdl_df]{fabletools::generate.mdl_df()}}).}

\item{...}{Other arguments passed to methods}
}
\description{
Simulates future paths from a dataset using a fitted model.
}
\examples{
melb_rain |>
  model(logistic = LOGISTIC(Wet ~ fourier(K = 5, period = "year"))) |>
  generate()
}
