% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet.R
\name{glance.BINNET}
\alias{glance.BINNET}
\title{Glance a BINNET model}
\usage{
\method{glance}{BINNET}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{Other arguments ignored}
}
\value{
A one row tibble summarising the model's fit.
}
\description{
Construct a single row summary of the BINNET model.
Contains the variance of residuals (\code{sigma2}).
}
\examples{
as_tsibble(airmiles) \%>\%
  model(nn = BINNET(box_cox(value, 0.15))) \%>\%
  glance()
}
