% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet.R
\name{residuals.BINNET}
\alias{residuals.BINNET}
\title{Extract residuals from a fable model.}
\usage{
\method{residuals}{BINNET}(object, ...)
}
\arguments{
\item{object}{Fitted model}

\item{...}{Other arguments ignored}
}
\description{
Extracts the residuals.
}
\examples{
melb_rain |>
  model(nn = BINNET(Wet ~ fourier(K = 1, period = "year"))) |>
  residuals()
}
