% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{residuals.LOGISTIC}
\alias{residuals.LOGISTIC}
\title{Extract residuals from a fable model.}
\usage{
\method{residuals}{LOGISTIC}(
  object,
  type = c("deviance", "innovation", "pearson", "working", "response", "partial"),
  ...
)
}
\arguments{
\item{object}{Fitted model}

\item{type}{the type of residuals which should be returned.
alternatives are: "deviance" (default), "pearson", "working", "response",
and "partial". Can be abbreviated.}

\item{...}{Other arguments ignored}
}
\description{
Extracts the residuals.
}
\examples{
melb_rain |>
  model(logistic = LOGISTIC(Wet ~ fourier(K = 5, period = "year"))) |>
  residuals(type = "deviance")
}
